Trying to clean up the error log
I know there's a lot of changes here, but it makes the error log smaller. There's a lot of Bronze Age script lurking around, and removing that is a big part of this. I also deleted files that aren't different from the vanilla versions and changed files to be closer to vanilla versions where the mod versions were outdated. Currently, some of my changes cause errors because I haven't fixed up all the events yet, but this is already so big that I wanted to send these out.