Patch 1.6 compatibility part 1
This part should cover the files that we already had which were changed in the latest patch, making our versions use new scripted effects, incorporate bugfixes, etc. Part 2 will be changes to the new files added in patch 1.6, so that they don't refer to cultures/religions/geography that we don't have.