(Hopefully) dealing with the clothes-less Almyran & Morfite Problem once and for all
So the problem was that the only clothing_gfx the Almyrans and Morfites had were locked behind a DLC that many of us probably don't have. That has been fixed.

Oh, and I may have also changed the Almyrans to be more Iranian, but that's besides the point. No longer will the Almyrans and Morfites be without clothing!